Revert "API IN CittyPuzzle (no cleanup) (#61)"
This reverts commit 934b220b2922556d3c575f79405606c3c672656d.

+++ CityPuzzle/Login And SingUp/SignUpPageLogic.cs
@@ -19,7 +19,7 @@ namespace CityPuzzle
                 UserName = userNameEntry,
                 Email = emailEntry,
                 Pass = tempUser.PassToHash(passEntry),
-                MaxQuestDistance = int.Parse(distEntry)
+                MaxQuestDistance = double.Parse(distEntry)
             };
 
             Sql.SaveUser(user);
